Linter: Move `path` to CLI-only part
