+++ javascript/packages/linter/src/cli/file-url.ts
@@ -1,0 +1,6 @@
+import { resolve } from "node:path"
+
+export function fileUrl(filePath: string): string {
+  const absolutePath = resolve(filePath)
+  return `file://${absolutePath}`
+}
